requirements.txt
Flask==2.3.3
google-cloud-bigquery==3.36.0
google-cloud-storage
pandas==2.3.2
pyarrow==21.0.0
functions-framework
psutil==5.9.8
gunicorn==21.2.0